fix(exchange):edit buttons were wild
