+++ apps/www/widgets/exchanges/card/index.tsx
@@ -1,5 +1,6 @@
 "use client";
 
+import { AppToastOptions } from ":components/toast";
 import { TRPC_React } from ":trpc/client";
 import type { Exchange } from "@1.modules/exchange.domain";
 import { Card } from "@1.modules/exchange.ui/Card/Card";
@@ -14,11 +15,19 @@ import {
 import { exchange_card } from "@1.modules/exchange.ui/Card/exchange_card";
 import { PROFILE_ROLES, type Profile } from "@1.modules/profile.domain";
 import { StudentAvatarMedia } from "@1.modules/profile.ui/avatar";
-import { Exchange as ExchangeIcon } from "@1.ui/react/icons";
+import { button } from "@1.ui/react/button/atom";
+import { Exchange as ExchangeIcon, Pen } from "@1.ui/react/icons";
 import { useTimeoutEffect } from "@react-hookz/web";
 import dynamic from "next/dynamic";
 import Link from "next/link";
-import { type PropsWithChildren } from "react";
+import {
+  useCallback,
+  useEffect,
+  useState,
+  type MouseEventHandler,
+  type PropsWithChildren,
+} from "react";
+import { toast } from "react-toastify";
 import { P, match } from "ts-pattern";
 import { Exchange_Actions } from "./actions";
 import { Exchange_Bookmark } from "./bookmark";
@@ -97,30 +106,39 @@ function Idle() {
   return (
     <Card_Idle>
       <Card.Header.Left>
-        <Link href={`/@${exchange.owner.profile.id}`}>
-          <StudentAvatarMedia
-            className="h-full items-center"
-            student={exchange.owner}
-          />
-        </Link>
-      </Card.Header.Left>
-      <Card.Header.Right>
-        <div className="flex flex-row items-center gap-4">
-          <div className="hidden items-center justify-between text-[#707070] md:flex">
-            <span className="whitespace-nowrap font-bold uppercase">
-              {exchange.category.name}
-            </span>
-            <ExchangeIcon
-              className={exchange_icon({
-                with_return: Boolean(exchange.return),
-              })}
+        <div className="flex h-full w-full justify-between gap-2">
+          <Link href={`/@${exchange.owner.profile.id}`}>
+            <StudentAvatarMedia
+              className="h-full items-center"
+              student={exchange.owner}
             />
-            <span className="whitespace-nowrap font-bold uppercase">
-              {match(exchange.return)
-                .with(null, () => "Sans échange")
-                .with(P._, (category) => category.name)
-                .exhaustive()}
-            </span>
+          </Link>
+          <div className="mr-2 flex items-end gap-2 md:hidden">
+            {is_yours ? <Edit_Buttons /> : null}
+          </div>
+        </div>
+      </Card.Header.Left>
+
+      <Card.Header.Right>
+        <div className="flex flex-col items-end ">
+          {is_yours ? <Edit_Buttons /> : null}
+          <div className="flex flex-row items-center gap-4">
+            <div className="hidden items-center justify-between text-[#707070] md:flex">
+              <span className="whitespace-nowrap font-bold uppercase">
+                {exchange.category.name}
+              </span>
+              <ExchangeIcon
+                className={exchange_icon({
+                  with_return: Boolean(exchange.return),
+                })}
+              />
+              <span className="whitespace-nowrap font-bold uppercase">
+                {match(exchange.return)
+                  .with(null, () => "Sans échange")
+                  .with(P._, (category) => category.name)
+                  .exhaustive()}
+              </span>
+            </div>
           </div>
         </div>
       </Card.Header.Right>
@@ -140,7 +158,6 @@ function Idle() {
         <Exchange_Actions />
       </Card.Footer.Center>
       <Card.Footer.Right>
-        {is_yours ? <Edit_Buttons /> : null}
         <Exchange_Share />
       </Card.Footer.Right>
     </Card_Idle>
@@ -149,8 +166,56 @@ function Idle() {
 
 export function Edit_Buttons() {
   return (
-    <>
+    <div className="flex gap-3 md:gap-1">
+      <Exchange_Modify_Button />
       <Exchange_Delete_Button />
+    </div>
+  );
+}
+
+function Exchange_Modify_Button() {
+  const exchange = useExchange();
+  const [showWarning, setShowWarning] = useState(false);
+
+  const alreadyPopulated = exchange.deals.length > 0;
+
+  const on_edit_click = useCallback<MouseEventHandler<HTMLAnchorElement>>(
+    (event) => {
+      if (!alreadyPopulated) return;
+      event.preventDefault();
+
+      setShowWarning(true);
+      setTimeout(() => {
+        setShowWarning(false);
+      }, 1_000 * 5);
+    },
+    [alreadyPopulated],
+  );
+
+  useEffect(() => {
+    if (showWarning) {
+      toast.info(
+        "Les échanges comprenant déjà des participants ne peuvent pas être édités",
+        AppToastOptions,
+      );
+    }
+  }, [showWarning]);
+  return (
+    <>
+      <Link
+        className={button({
+          intent: "light",
+          size: "xs",
+          state: "ghost",
+          className: "box-content h-4 text-gray-500 md:h-7 md:p-2",
+        })}
+        href={alreadyPopulated ? "" : `/@~/exchanges/${exchange.id}/edit`}
+        onClick={on_edit_click}
+      >
+        <Pen className="h-4 md:h-7" />
+      </Link>
     </>
   );
 }
+
+export default Exchange_Modify_Button;
