+++ packages/@1/modules/exchange/ui/src/Card/Card_Idle.tsx
@@ -1,22 +1,14 @@
 //
 
 import { AvatarMedia } from "@1.ui/react/avatar";
-import { button } from "@1.ui/react/button/atom";
-import { ExclamationMark, Pen, School, Share } from "@1.ui/react/icons";
+import { ExclamationMark, School, Share } from "@1.ui/react/icons";
 import { ActionItem, Menu } from "@1.ui/react/menu";
-import Link from "next/link";
-import {
-  useCallback,
-  useEffect,
-  useState,
-  type MouseEventHandler,
-  type PropsWithChildren,
-} from "react";
+import { type PropsWithChildren } from "react";
 import { tv } from "tailwind-variants";
 import { Card } from "./Card";
 import { Publish_Date } from "./Date";
 import { InfoBar } from "./InfoBar";
-import { useExchange, useExchangeMeta } from "./context";
+import { useExchange } from "./context";
 import { exchange_card } from "./exchange_card";
 //
 
@@ -27,8 +19,6 @@ export function Card_Idle({ children }: PropsWithChildren) {
     type: exchange.type,
   });
   const { title, description } = exchange;
-  const { is_yours } = useExchangeMeta();
-  const { withEdit, classic } = style();
 
   return (
     <div id={exchange.id} className={base()}>
@@ -43,12 +33,15 @@ export function Card_Idle({ children }: PropsWithChildren) {
             </AvatarMedia>
           </Card.Header.Left.Renderer>
           <Card.Header.Center.Renderer childs={children}>
-            <div>
-              <figure className="flex flex-col items-end md:items-center ">
+            <div
+              className="
+            flex h-full items-center justify-center"
+            >
+              <figure className="flex flex-col items-end md:items-center">
                 <div className="text-xl font-bold text-primary">
                   {`${exchange.deals.length} / ${exchange.places}`}
                 </div>
-                <figcaption className="text-center text-sm">
+                <figcaption className="text-center text-gray-500">
                   participant{exchange.places > 1 ? "s" : ""}
                 </figcaption>
               </figure>
@@ -81,7 +74,6 @@ export function Card_Idle({ children }: PropsWithChildren) {
         </article>
         <div className="flex justify-between">
           <Publish_Date />
-          {is_yours ? <Edit_Buttons /> : null}
         </div>
       </div>
       <footer className={footer()}>
@@ -95,7 +87,7 @@ export function Card_Idle({ children }: PropsWithChildren) {
             ...
           </Card.Footer.Center.Renderer>
         </div>
-        <div className={is_yours ? withEdit() : classic()}>
+        <div className={style()}>
           <Card.Footer.Right.Renderer childs={children}>
             <Share className="size-5" />
           </Card.Footer.Right.Renderer>
@@ -104,62 +96,6 @@ export function Card_Idle({ children }: PropsWithChildren) {
     </div>
   );
 }
-
-export function Edit_Buttons() {
-  const exchange = useExchange();
-  const [showWarning, setShowWarning] = useState(false);
-
-  const alreadyPopulated = exchange.deals.length > 0;
-
-  const on_edit_click = useCallback<MouseEventHandler<HTMLAnchorElement>>(
-    (event) => {
-      if (!alreadyPopulated) return;
-      event.preventDefault();
-
-      setShowWarning(true);
-      setTimeout(() => {
-        setShowWarning(false);
-      }, 1_000 * 5);
-    },
-    [alreadyPopulated],
-  );
-
-  useEffect(() => {
-    if (showWarning) {
-      <InfoBox message="Les échanges comprenant déjà des participants ne peuvent pas être édités" />;
-    }
-  }, [showWarning]);
-  return (
-    <>
-      <Link
-        className={button({
-          intent: "light",
-          size: "sm",
-          state: "ghost",
-          className: "mt-8 box-content h-4 py-2 text-black",
-        })}
-        href={alreadyPopulated ? "" : `/@~/exchanges/${exchange.id}/edit`}
-        onClick={on_edit_click}
-      >
-        <Pen className="h-4" />
-      </Link>
-    </>
-  );
-}
-
-interface InfoBoxProps {
-  message: string;
-}
-
-export function InfoBox({ message }: InfoBoxProps) {
-  return (
-    <div className="fixed left-0 top-0 z-[51] ml-0 flex w-full items-center justify-center bg-red-500 p-6 text-white md:bottom-2 md:top-auto md:w-[25%] md:rounded-e-2xl">
-      {message}
-    </div>
-  );
-}
-
-export default InfoBox;
 
 export function ExchangeMenu({ exchange_id }: { exchange_id: string }) {
   const href = `/exchanges/${exchange_id}`;
@@ -177,9 +113,5 @@ export function ExchangeMenu({ exchange_id }: { exchange_id: string }) {
 }
 
 const style = tv({
-  base: "",
-  slots: {
-    withEdit: "col-start-6 flex items-center gap-2 justify-self-end",
-    classic: "col-start-6 justify-self-end",
-  },
+  base: "col-start-6 flex items-center justify-self-end",
 });
